webapp: split the graph logic based on the graph subtypes
- TokenizedText not rendering properly now
