+++ frontend/src/components/Graph.tsx
@@ -1,776 +1,71 @@
 import React from 'react';
 import { FigureConfig, VisualizationData } from '../types';
 import { getGraphType } from '../utils/graphTypes';
-import { buildLanguageLabelMap, getDisplayLanguageLabel, getDisplayLanguagePairLabel } from '../utils/languageLabels';
-import { decodeSentence } from '../utils/tokenDecoder';
-import {
-  BarChart,
-  Bar,
-  Scatter,
-  XAxis,
-  YAxis,
-  CartesianGrid,
-  Tooltip,
-  Legend,
-  ResponsiveContainer,
-  ComposedChart,
-  Line,
-} from 'recharts';
+import BarGraph from './graphs/BarGraph';
+import ScatterGraph from './graphs/ScatterGraph';
+import MetricTable from './graphs/MetricTable';
+import TokenizedText from './graphs/TokenizedText';
 import './Graph.css';
-
-const GRAPH_COLORS = [
-  '#8884d8',
-  '#82ca9d',
-  '#ffc658',
-  '#ff7c7c',
-  '#8dd1e1',
-  '#d084d0',
-  '#ffa500',
-  '#00ff00',
-];
-
-const getColorForMetric = (index: number): string => GRAPH_COLORS[index % GRAPH_COLORS.length];
 
 interface GraphProps {
   config: FigureConfig;
   data: VisualizationData;
 }
 
-const ScatterTooltip: React.FC<{
-  active?: boolean;
-  payload?: any[];
-  metricX: string;
-  metricY: string;
-  formatLanguage?: (label: string, isPair: boolean) => string;
-}> = ({
-  active,
-  payload,
-  metricX,
-  metricY,
-  formatLanguage,
-}) => {
-  if (!active || !payload || payload.length === 0) return null;
-
-  const pt = payload[0]?.payload || {};
-  const hasLanguagePair = pt.languagePair !== undefined && pt.languagePair !== null;
-  const rawLanguageLabel = hasLanguagePair ? pt.languagePair : pt.language;
-  const languageLabel =
-    rawLanguageLabel !== undefined && formatLanguage
-      ? formatLanguage(String(rawLanguageLabel), hasLanguagePair)
-      : rawLanguageLabel;
-  const languageTitle = hasLanguagePair ? 'Language pair' : 'Language';
-
-   const clusterSize = typeof pt.clusterSize === 'number' ? pt.clusterSize : undefined;
-
-  const xVal = pt[metricX];
-  const yVal = pt[metricY];
-
-  const formatVal = (v: any) => (typeof v === 'number' ? v.toFixed(2) : v);
-
-  return (
-    <div style={{ backgroundColor: '#fff', border: '1px solid #ccc', padding: '5px' }}>
-      <p><strong>Tokenizer: {pt.tokenizer ?? 'N/A'}</strong></p>
-      {languageLabel !== undefined && (
-        <p>{languageTitle}: {languageLabel}</p>
-      )}
-      {clusterSize !== undefined && clusterSize > 1 && (
-        <p>Points in cluster: {clusterSize}</p>
-      )}
-      <p>{metricX}: {formatVal(xVal)}</p>
-      <p>{metricY}: {formatVal(yVal)}</p>
-    </div>
-  );
-};
-
-const MAX_SCATTER_POINTS_PER_GROUP = 2000;
-
-const downsampleGroupPoints = (
-  points: any[],
-  metricX: string,
-  metricY: string,
-  maxPointsPerGroup: number,
-): any[] => {
-  if (!Array.isArray(points) || points.length <= maxPointsPerGroup) {
-    return points;
-  }
-
-  let minX = Infinity;
-  let maxX = -Infinity;
-  let minY = Infinity;
-  let maxY = -Infinity;
-
-  for (const pt of points) {
-    const x = pt[metricX];
-    const y = pt[metricY];
-    if (typeof x !== 'number' || typeof y !== 'number' || Number.isNaN(x) || Number.isNaN(y)) {
-      continue;
-    }
-    if (x < minX) minX = x;
-    if (x > maxX) maxX = x;
-    if (y < minY) minY = y;
-    if (y > maxY) maxY = y;
-  }
-
-  if (!Number.isFinite(minX) || !Number.isFinite(maxX) || !Number.isFinite(minY) || !Number.isFinite(maxY)) {
-    return points;
-  }
-
-  const rangeX = maxX - minX || 1;
-  const rangeY = maxY - minY || 1;
-
-  const gridSize = Math.ceil(Math.sqrt(maxPointsPerGroup));
-  type Cell = { sumX: number; sumY: number; count: number; template: any };
-  const cells: Map<string, Cell> = new Map();
-
-  for (const pt of points) {
-    const x = pt[metricX];
-    const y = pt[metricY];
-    if (typeof x !== 'number' || typeof y !== 'number' || Number.isNaN(x) || Number.isNaN(y)) {
-      continue;
-    }
-
-    const gxRaw = ((x - minX) / rangeX) * gridSize;
-    const gyRaw = ((y - minY) / rangeY) * gridSize;
-    const gx = Math.min(gridSize - 1, Math.max(0, Math.floor(gxRaw)));
-    const gy = Math.min(gridSize - 1, Math.max(0, Math.floor(gyRaw)));
-    const key = `${gx}:${gy}`;
-
-    let cell = cells.get(key);
-    if (!cell) {
-      cell = { sumX: 0, sumY: 0, count: 0, template: pt };
-      cells.set(key, cell);
-    }
-
-    cell.sumX += x;
-    cell.sumY += y;
-    cell.count += 1;
-  }
-
-  const result: any[] = [];
-
-  cells.forEach((cell) => {
-    if (cell.count === 0) return;
-    const centroidX = cell.sumX / cell.count;
-    const centroidY = cell.sumY / cell.count;
-    const centroid: any = {
-      ...cell.template,
-      [metricX]: centroidX,
-      [metricY]: centroidY,
-    };
-    if (cell.count > 1) {
-      centroid.clusterSize = cell.count;
-    }
-    result.push(centroid);
-  });
-
-  return result;
-};
-
 const Graph: React.FC<GraphProps> = ({ config, data }) => {
   const graphType = getGraphType(config.typeId);
 
-  type RowSortState = { columnIndex: number | null; direction: 'asc' | 'desc' };
-  type ColumnSortState = { rowIndex: number | null; direction: 'asc' | 'desc' };
-  const [rowSort, setRowSort] = React.useState<RowSortState>({ columnIndex: null, direction: 'asc' });
-  const [columnSort, setColumnSort] = React.useState<ColumnSortState>({ rowIndex: null, direction: 'asc' });
-  const [hoveredToken, setHoveredToken] = React.useState<string | null>(null);
-  const [hoveredSentenceIdx, setHoveredSentenceIdx] = React.useState<number | null>(null);
-
-  const getChartData = (): any[] | any => {
-    if (!graphType) return [];
-
-    try {
-      return graphType.transform(data, {
-        metrics: config.metrics,
-        tokenizers: config.tokenizers,
-        languages: config.languages,
-      });
-    } catch (error) {
-      console.error('Error transforming chart data:', error);
-      return [];
-    }
-  };
-
-  const chartData = getChartData();
-
-  const allLanguages = data.metadata?.languages || [];
-  const languageLabelMap = React.useMemo(
-    () => buildLanguageLabelMap(allLanguages),
-    [allLanguages],
-  );
-
-  const renderChart = () => {
-    if (!graphType) {
-      return <div className="error">Graph type not found: {config.typeId}</div>;
-    }
-
-    const metrics = config.metrics;
-
-    // Tokenized-text type doesn't use chartData at all — render directly
-    if (config.typeId === 'tokenized-text') {
-      return renderTokenizedText();
-    }
-
-    // Check if we have valid data
-    const hasData = (Array.isArray(chartData) && chartData.length > 0) || 
-                    (chartData && typeof chartData === 'object' && (chartData.data || chartData.error));
-
-    if (!hasData) {
-      let errorMsg = 'No data available for this configuration';
-      if (chartData?.error) {
-        errorMsg = chartData.error;
-      }
-      
-      return <div className="no-data">{errorMsg}</div>;
-    }
-
-    if (config.typeId === 'metric-pair-correlation') {
-      return renderScatterChart(metrics);
-    } else if (config.typeId === 'metric-table') {
-      return renderMetricTable();
-    } else {
-      return renderBarChart(metrics);
-    }
-  };
-
-  const renderTokenizedText = () => {
-    const tokenizations: Record<string, any> | undefined = data.metadata?.tokenizations;
-    if (!tokenizations) {
-      return <div className="no-data">No tokenizations available for this dataset</div>;
-    }
-
-    const selectedTokenizers = config.tokenizers && config.tokenizers.length > 0
-      ? config.tokenizers
-      : Object.keys(tokenizations);
-    const selectedLanguages = config.languages && config.languages.length > 0
-      ? config.languages
-      : [];
-
-    if (selectedLanguages.length === 0) {
-      return <div className="no-data">Select at least one language to display tokenized text</div>;
-    }
-
-    const panels: React.ReactNode[] = [];
-    for (const tokenizer of selectedTokenizers) {
-      const tokData = tokenizations[tokenizer];
-      if (!tokData) continue;
-      for (const lang of selectedLanguages) {
-        const sentences: string[][] | undefined = tokData[lang];
-        if (!sentences) continue;
-        const from = config.sentenceRange ? config.sentenceRange[0] - 1 : 0;
-        const to = config.sentenceRange ? config.sentenceRange[1] : sentences.length;
-        const slice = sentences.slice(Math.max(0, from), to);
-        panels.push(
-          <div key={`${tokenizer}__${lang}`} className="tokenized-text-panel">
-            <div className="tokenized-text-header">
-              <span className="tokenized-text-tokenizer">{tokenizer}</span>
-              <span className="tokenized-text-lang">{lang}</span>
-            </div>
-            <div className="tokenized-text-body">
-              {slice.map((tokens, i) => (
-                <div
-                  key={from + i}
-                  className={['tokenized-text-sentence', hoveredSentenceIdx === from + i ? 'tokenized-text-sentence--active' : ''].filter(Boolean).join(' ')}
-                  onMouseEnter={() => setHoveredSentenceIdx(from + i)}
-                  onMouseLeave={() => setHoveredSentenceIdx(null)}
-                >
-                  <span className="tokenized-text-idx">{from + i + 1}.</span>{' '}
-                  {decodeSentence(tokens).map((dt, j, arr) => (
-                    <React.Fragment key={j}>
-                      <span
-                        className={[
-                          'tokenized-text-token',
-                          dt.isByteLevel ? 'tokenized-text-token--byte' : '',
-                          dt.decoded === '' ? 'tokenized-text-token--fragment' : '',
-                          hoveredToken === dt.raw ? 'tokenized-text-token--active' : '',
-                        ].filter(Boolean).join(' ')}
-                        title={dt.isByteLevel && dt.decoded !== dt.raw ? `raw: ${dt.raw}` : undefined}
-                        onMouseEnter={() => setHoveredToken(dt.raw)}
-                        onMouseLeave={() => setHoveredToken(null)}
-                      >{dt.decoded !== '' ? dt.decoded : <span className="tokenized-text-fragment-placeholder">·</span>}</span>
-                      {j < arr.length - 1 && arr[j + 1].decoded !== '' && ' '}
-                    </React.Fragment>
-                  ))}
-                </div>
-              ))}
-            </div>
-          </div>
-        );
-      }
-    }
-
-    if (panels.length === 0) {
-      return <div className="no-data">No data for the selected tokenizer/language combination</div>;
-    }
-
-    return <div className="tokenized-text-container">{panels}</div>;
-  };
-
-  const renderBarChart = (metrics: string[]) => {
+  // Tokenized-text is data-independent — render directly.
+  if (config.typeId === 'tokenized-text') {
     return (
-      <ResponsiveContainer width="100%" height={400}>
-        <BarChart data={chartData} margin={{ top: 20, right: 30, left: 60, bottom: 100 }}>
-          <CartesianGrid strokeDasharray="3 3" />
-          <XAxis 
-            dataKey="name" 
-            angle={-45} 
-            textAnchor="end" 
-            height={120}
-            interval={0}
-          />
-          <YAxis 
-            type="number" 
-            domain={[0, 16]} 
-            ticks={[0, 2, 4, 6, 8, 10, 12, 14, 16]}
-            label={{ value: 'Ranking (1=Best)', angle: -90, position: 'insideLeft', offset: 10 }} 
-          />
-          <Tooltip />
-          <Legend />
-          {metrics.map((metric, index) => (
-              <Bar
-                key={metric}
-                dataKey={metric}
-                fill={getColorForMetric(index)}
-                name={metric}
-                isAnimationActive={false}
-              />
-          ))}
-        </BarChart>
-      </ResponsiveContainer>
-    );
-  };
-
-  const renderScatterChart = (metrics: string[]) => {
-    const metricX = metrics[0];
-    const metricY = metrics[1];
-
-    const groupBy = config.groupBy || 'tokenizer';
-
-    const getLanguageInfoForLabel = (label: string): any => {
-      const parts = label.split('_');
-      const base = parts.length >= 3 ? parts.slice(0, parts.length - 2).join('_') : label;
-      const glottocode = parts.length >= 3 ? parts[parts.length - 1] : undefined;
-      const root: any = data?.metadata?.languagesInfo || {};
-      const direct = root.languages && root.languages[base] ? root.languages[base]
-        : (root[base] ? root[base] : null);
-      if (direct) return direct;
-      if (glottocode) {
-        const entries: Array<{ key: string; info: any }> = root.languages && typeof root.languages === 'object'
-          ? Object.keys(root.languages).map((k) => ({ key: k, info: root.languages[k] }))
-          : Object.keys(root).map((k) => ({ key: k, info: root[k] }));
-        for (const { info } of entries) {
-          const gc = info?.glottocodes;
-          if (typeof gc === 'string' && gc === glottocode) return info;
-          if (Array.isArray(gc) && gc.includes(glottocode)) return info;
-          if (gc && typeof gc === 'object' && Object.keys(gc).includes(glottocode)) return info;
-        }
-      }
-      return null;
-    };
-
-    const getFamilyForLanguage = (label: string): string => {
-      const info = getLanguageInfoForLabel(label);
-      if (!info) return 'unknown';
-      const fam = info.families;
-      if (Array.isArray(fam)) return fam[0] || 'unknown';
-      if (fam && typeof fam === 'object') {
-        const keys = Object.keys(fam);
-        return keys[0] || 'unknown';
-      }
-      if (typeof fam === 'string') return fam || 'unknown';
-      return 'unknown';
-    };
-
-    const groupKeyForPoint = (pt: any): string => {
-      if (groupBy === 'tokenizer') return pt.tokenizer || 'unknown';
-      if (groupBy === 'language') return pt.language || 'unknown';
-      if (groupBy === 'family') return getFamilyForLanguage(pt.language || '');
-      return 'unknown';
-    };
-
-    const isValidPoint = (pt: any): boolean => {
-      const x = pt[metricX];
-      const y = pt[metricY];
-      return typeof x === 'number' && typeof y === 'number' && !Number.isNaN(x) && !Number.isNaN(y) && isFinite(x) && isFinite(y);
-    };
-
-    const allPoints: any[] = (Array.isArray(chartData) ? chartData : []).filter(isValidPoint);
-
-    // Partition chartData into groups (only valid points)
-    const groupsMap: Map<string, any[]> = new Map();
-    allPoints.forEach((pt) => {
-      const key = groupKeyForPoint(pt);
-      if (!groupsMap.has(key)) groupsMap.set(key, []);
-      groupsMap.get(key)!.push(pt);
-    });
-    const groupNames = Array.from(groupsMap.keys());
-
-    const computeTrend = (pts: any[]): { m: number; b: number; minX: number; maxX: number } | null => {
-      if (!pts || pts.length < 2) return null;
-      let sumX = 0, sumY = 0, sumXY = 0, sumXX = 0;
-      let minX = Infinity, maxX = -Infinity;
-      for (const p of pts) {
-        const x = p[metricX];
-        const y = p[metricY];
-        sumX += x;
-        sumY += y;
-        sumXY += x * y;
-        sumXX += x * x;
-        if (x < minX) minX = x;
-        if (x > maxX) maxX = x;
-      }
-      const n = pts.length;
-      const denom = (n * sumXX - sumX * sumX);
-      if (denom === 0) return null;
-      const m = (n * sumXY - sumX * sumY) / denom;
-      const b = (sumY - m * sumX) / n;
-      return { m, b, minX, maxX };
-    };
-
-    const trendlineMode: 'none' | 'global' | 'groups' =
-      (config as any).trendlineMode || (config.showTrendline ? 'global' : 'none');
-
-    const buildTrendData = (trend: { m: number; b: number; minX: number; maxX: number }) => ([
-      { [metricX]: trend.minX, [metricY]: trend.m * trend.minX + trend.b },
-      { [metricX]: trend.maxX, [metricY]: trend.m * trend.maxX + trend.b },
-    ]);
-
-    type TrendLineDef = { key: string; name: string; color: string; data: any[] };
-    const trendLines: TrendLineDef[] = [];
-
-    if (trendlineMode === 'global') {
-      const trend = computeTrend(allPoints);
-      if (trend) {
-        trendLines.push({
-          key: 'global-trend',
-          name: 'Trend (global)',
-          color: '#444',
-          data: buildTrendData(trend),
-        });
-      }
-    } else if (trendlineMode === 'groups') {
-      groupNames.forEach((name, idx) => {
-        const groupPts = groupsMap.get(name) || [];
-        const trend = computeTrend(groupPts);
-        if (trend) {
-          trendLines.push({
-            key: `trend-${name}`,
-            name: `${name} trend`,
-            color: getColorForMetric(idx),
-            data: buildTrendData(trend),
-          });
-        }
-      });
-    }
-
-    // Downsample large groups to avoid rendering too many points
-    const displayGroupsMap: Map<string, any[]> = new Map();
-    groupNames.forEach((name) => {
-      const groupPts = groupsMap.get(name) || [];
-      const displayPts = downsampleGroupPoints(groupPts, metricX, metricY, MAX_SCATTER_POINTS_PER_GROUP);
-      displayGroupsMap.set(name, displayPts);
-    });
-
-    return (
-      <ResponsiveContainer width="100%" height={420}>
-        <ComposedChart margin={{ top: 20, right: 30, left: 60, bottom: 90 }}>
-          <CartesianGrid strokeDasharray="3 3" />
-          <XAxis 
-            type="number" 
-            dataKey={metricX} 
-            name={metricX}
-            label={{ value: metricX, position: 'insideBottomRight', offset: -10 }}
-          />
-          <YAxis 
-            type="number" 
-            dataKey={metricY} 
-            name={metricY}
-            label={{ value: metricY, angle: -90, position: 'insideLeft', offset: 10 }}
-          />
-          <Tooltip 
-            cursor={{ strokeDasharray: '3 3' }}
-              content={(
-                <ScatterTooltip
-                  metricX={metricX}
-                  metricY={metricY}
-                  formatLanguage={(label, isPair) =>
-                    isPair
-                      ? getDisplayLanguagePairLabel(label, languageLabelMap)
-                      : getDisplayLanguageLabel(label, languageLabelMap)
-                  }
-                />
-              )}
-          />
-          <Legend verticalAlign="bottom" align="center" wrapperStyle={{ paddingTop: 10 }} />
-          {trendLines.map((t) => (
-            <Line
-              key={t.key}
-              type="linear"
-              data={t.data}
-              dataKey={metricY}
-              name={t.name}
-              stroke={t.color}
-              strokeDasharray="4 2"
-              dot={false}
-              isAnimationActive={false}
-            />
-          ))}
-          {groupNames.map((name, idx) => (
-            <Scatter
-              key={name}
-              name={name}
-              data={displayGroupsMap.get(name)!}
-              dataKey={metricY}
-              fill={getColorForMetric(idx)}
-              isAnimationActive={false}
-            />
-          ))}
-        </ComposedChart>
-      </ResponsiveContainer>
-    );
-  };
-
-  const renderMetricTable = () => {
-    const tableData = chartData as any;
-    
-    // Check for error message
-    if (tableData.error) {
-      return <div className="no-data"><strong>Error:</strong> {tableData.error}</div>;
-    }
-    
-    if (!tableData.rows || !tableData.columns || !tableData.data) {
-      return <div className="no-data">Invalid table data structure</div>;
-    }
-
-    if (tableData.data.length === 0) {
-      return <div className="no-data">No data available for this configuration</div>;
-    }
-
-    const is3DMetricTable =
-      config.typeId === 'metric-table' &&
-      Array.isArray(config.metrics) &&
-      config.metrics.length === 1 &&
-      (data as any).metrics?.[config.metrics[0]]?.shape &&
-      (data as any).metrics[config.metrics[0]].shape.length === 3;
-
-    const renderColumnLabel = (col: string) => {
-      if (is3DMetricTable && col.includes('-')) {
-        const [lang1, lang2] = col.split('-', 2);
-        return (
-          <span className="metric-table-col-label">
-            <span>{getDisplayLanguageLabel(lang1, languageLabelMap)}</span>
-            <br />
-            <span>{getDisplayLanguageLabel(lang2, languageLabelMap)}</span>
-          </span>
-        );
-      }
-      return getDisplayLanguageLabel(col, languageLabelMap);
-    };
-
-    const handleColumnHeaderClick = (colIdx: number) => {
-      setRowSort((prev) => {
-        if (prev.columnIndex === colIdx) {
-          return { columnIndex: colIdx, direction: prev.direction === 'asc' ? 'desc' : 'asc' };
-        }
-        return { columnIndex: colIdx, direction: 'asc' };
-      });
-    };
-
-    const handleRowHeaderClick = (rowIdx: number) => {
-      setColumnSort((prev) => {
-        if (prev.rowIndex === rowIdx) {
-          return { rowIndex: rowIdx, direction: prev.direction === 'asc' ? 'desc' : 'asc' };
-        }
-        return { rowIndex: rowIdx, direction: 'asc' };
-      });
-    };
-
-    const sortedRowIndices: number[] = tableData.data.map((_: any, idx: number) => idx);
-    if (rowSort.columnIndex !== null && rowSort.columnIndex >= 0 && rowSort.columnIndex < tableData.columns.length) {
-      const colIdx = rowSort.columnIndex;
-      sortedRowIndices.sort((a, b) => {
-        const cellA = tableData.data[a][colIdx];
-        const cellB = tableData.data[b][colIdx];
-        const vaRaw = cellA ? cellA.value : undefined;
-        const vbRaw = cellB ? cellB.value : undefined;
-        const va = vaRaw === undefined || vaRaw === null || Number.isNaN(Number(vaRaw)) ? Infinity : Number(vaRaw);
-        const vb = vbRaw === undefined || vbRaw === null || Number.isNaN(Number(vbRaw)) ? Infinity : Number(vbRaw);
-        if (rowSort.direction === 'asc') return va - vb;
-        return vb - va;
-      });
-    }
-
-    const sortedColumnIndices: number[] = tableData.columns.map((_: any, idx: number) => idx);
-    if (columnSort.rowIndex !== null && columnSort.rowIndex >= 0 && columnSort.rowIndex < tableData.data.length) {
-      const rowIdx = columnSort.rowIndex;
-      const row = tableData.data[rowIdx];
-      sortedColumnIndices.sort((a, b) => {
-        const cellA = row[a];
-        const cellB = row[b];
-        const vaRaw = cellA ? cellA.value : undefined;
-        const vbRaw = cellB ? cellB.value : undefined;
-        const va = vaRaw === undefined || vaRaw === null || Number.isNaN(Number(vaRaw)) ? Infinity : Number(vaRaw);
-        const vb = vbRaw === undefined || vbRaw === null || Number.isNaN(Number(vbRaw)) ? Infinity : Number(vbRaw);
-        if (columnSort.direction === 'asc') return va - vb;
-        return vb - va;
-      });
-    }
-
-    let minValue = Infinity;
-    let maxValue = -Infinity;
-
-    sortedRowIndices.forEach((rowIdx: number) => {
-      const row = tableData.data[rowIdx];
-      sortedColumnIndices.forEach((colIdx: number) => {
-        const cell = row[colIdx];
-        if (!cell || cell.value === undefined || cell.value === null) {
-          return;
-        }
-        const v = Number(cell.value);
-        if (!Number.isFinite(v)) {
-          return;
-        }
-        if (v < minValue) minValue = v;
-        if (v > maxValue) maxValue = v;
-      });
-    });
-
-    const hasValueRange = Number.isFinite(minValue) && Number.isFinite(maxValue);
-
-    const getCellBackground = (raw: any): React.CSSProperties['backgroundColor'] => {
-      if (!hasValueRange) return undefined;
-      const v = Number(raw);
-      if (!Number.isFinite(v)) return undefined;
-
-      if (minValue === maxValue) {
-        return 'rgba(76, 175, 80, 0.25)';
-      }
-
-      const t = (v - minValue) / (maxValue - minValue);
-      const clamped = Math.max(0, Math.min(1, t));
-
-      const start = { r: 232, g: 246, b: 231 }; // #e8f6e7 very light green
-      const end = { r: 0, g: 120, b: 50 };     // dark green
-
-      const r = Math.round(start.r + (end.r - start.r) * clamped);
-      const g = Math.round(start.g + (end.g - start.g) * clamped);
-      const b = Math.round(start.b + (end.b - start.b) * clamped);
-
-      return `rgb(${r}, ${g}, ${b})`;
-    };
-
-    const getCellTextColor = (raw: any): React.CSSProperties['color'] => {
-      if (!hasValueRange || minValue === maxValue) return undefined;
-      const v = Number(raw);
-      if (!Number.isFinite(v)) return undefined;
-
-      const t = (v - minValue) / (maxValue - minValue);
-      if (t >= 0.65) {
-        return '#ffffff';
-      }
-      return '#2c3e50';
-    };
-
-    const formatLegendValue = (v: number): string => {
-      if (!Number.isFinite(v)) return '';
-      const abs = Math.abs(v);
-      if ((abs >= 1000 && abs < Infinity) || (abs > 0 && abs < 0.01)) {
-        return v.toExponential(2);
-      }
-      return v.toFixed(3);
-    };
-
-    return (
-      <div className="metric-table-container">
-        <table className="metric-table">
-          <thead>
-            <tr>
-              <th className="table-row-header">{tableData.rowHeader || 'Tokenizer'}</th>
-              {sortedColumnIndices.map((colIdx: number) => {
-                const col = tableData.columns[colIdx];
-                const isActive = rowSort.columnIndex === colIdx;
-                return (
-                <th
-                  key={colIdx}
-                  className="table-column-header"
-                  onClick={() => handleColumnHeaderClick(colIdx)}
-                  style={{ cursor: 'pointer' }}
-                >
-                  {renderColumnLabel(col)}
-                  {isActive && (
-                    <span style={{ marginLeft: '4px' }}>
-                      {rowSort.direction === 'asc' ? '▲' : '▼'}
-                    </span>
-                  )}
-                </th>
-                );
-              })}
-            </tr>
-          </thead>
-          <tbody>
-            {sortedRowIndices.map((rowIdx: number) => {
-              const row = tableData.data[rowIdx];
-              return (
-                <tr key={rowIdx}>
-                  <td
-                    className="table-row-label"
-                    onClick={() => handleRowHeaderClick(rowIdx)}
-                    style={{ cursor: 'pointer' }}
-                  >
-                    {tableData.rows[rowIdx]}
-                    {columnSort.rowIndex === rowIdx && (
-                      <span style={{ marginLeft: '4px' }}>
-                        {columnSort.direction === 'asc' ? '▲' : '▼'}
-                      </span>
-                    )}
-                  </td>
-                  {sortedColumnIndices.map((colIdx: number) => {
-                    const cell = row[colIdx];
-                    const rawValue = cell && cell.value !== undefined ? cell.value : undefined;
-                    const backgroundColor = getCellBackground(rawValue);
-                    const color = getCellTextColor(rawValue);
-                    return (
-                    <td
-                      key={colIdx}
-                      className="table-cell"
-                      title={`Value: ${rawValue !== undefined ? rawValue : 'N/A'}`}
-                      style={{ backgroundColor, color }}
-                    >
-                      {cell ? cell.formatted : 'N/A'}
-                    </td>
-                    );
-                  })}
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
-        {hasValueRange && (
-          <div className="metric-table-legend">
-            <div className="metric-table-legend-title">
-              Cell color shows relative value (lower -&gt; higher)
-            </div>
-            <div className="metric-table-legend-bar-wrapper">
-              <div className="metric-table-legend-bar" />
-            </div>
-            <div className="metric-table-legend-labels">
-              <span>{formatLegendValue(minValue)}</span>
-              <span>{formatLegendValue((minValue + maxValue) / 2)}</span>
-              <span>{formatLegendValue(maxValue)}</span>
-            </div>
-          </div>
-        )}
+      <div className="graph" id={`graph-${config.id}`}>
+        <TokenizedText config={config} data={data} />
       </div>
     );
+  }
+
+  if (!graphType) {
+    return <div className="graph error">Graph type not found: {config.typeId}</div>;
+  }
+
+  let chartData: any;
+  try {
+    chartData = graphType.transform(data, {
+      metrics: config.metrics,
+      tokenizers: config.tokenizers,
+      languages: config.languages,
+    });
+  } catch (err) {
+    console.error('Error transforming chart data:', err);
+    return <div className="graph no-data">Error computing chart data</div>;
+  }
+
+  const hasData =
+    (Array.isArray(chartData) && chartData.length > 0) ||
+    (chartData && typeof chartData === 'object' && (chartData.data || chartData.error));
+
+  if (!hasData) {
+    const msg = chartData?.error || 'No data available for this configuration';
+    return (
+      <div className="graph" id={`graph-${config.id}`}>
+        <div className="no-data">{msg}</div>
+      </div>
+    );
+  }
+
+  const renderContent = () => {
+    if (config.typeId === 'metric-pair-correlation') {
+      return <ScatterGraph config={config} data={data} chartData={chartData} />;
+    }
+    if (config.typeId === 'metric-table') {
+      return <MetricTable config={config} data={data} tableData={chartData} />;
+    }
+    return <BarGraph config={config} data={data} chartData={chartData} />;
   };
 
   return (
     <div className="graph" id={`graph-${config.id}`}>
-      {config.typeId === 'tokenized-text' || (Array.isArray(chartData) ? chartData.length > 0 : !!chartData) ? (
-        renderChart()
-      ) : (
-        <div className="no-data">No data available for this configuration</div>
-      )}
+      {renderContent()}
     </div>
   );
 };
